Fix bug 3044545: Define X, Y, Z 16-bit register pairs
Atmel doc mentions "X,Y,Z as the Indirect Address Registers", made of
course of 8-bit register pairs as you have mentioned. Avrasm2 addendum also
indicates those are valid built-in register indications. I personally find
them more intuitive (as movsw is 16bit), but it's only a matter of taste,
of course. x,y,z are technically aliases of the 8-bit register in a sense
that movsw encodes the lower byte register (r26, r28, r30), but that's only
internal to the assembler.