Add hash table implementation
This is the first step towards replacing all the linked lists used for
labels, .defs, etc. The linked lists are a poor choice for what they're
used for. E.g., a file with n labels takes O(n^2) time to register in
the linked lists because it has to linear search the linked list for
duplicates.

Currently, the hash table supports only string keys and string/integer
values, but these seems to be sufficient for our needs.

I've manually tested and benchmarked the hash table, and it seems
alright with adequate performance. There are a few easy optimizations
that could be made if needed (e.g., caching overflow buckets and storing
the top byte(s) of hashes to avoid strcmp with high probability for most
operations). Also, valgrind reports no memory leaks.

The declarations for hash.c are in avra2.h, since I intend to rewrite
most of avra.c and avra.h.